Organize app into domain subpackages
