+++ docker-compose.yml
@@ -22,6 +22,6 @@ services:
     environment:
       CELERY_BROKER_URL: redis://redis:6379/0
       CELERY_RESULT_BACKEND: redis://redis:6379/0
-    command: celery -A app.tasks worker --loglevel=info
+    command: celery -A app.execution.tasks worker --loglevel=info
     depends_on:
       - redis
